Open windows in client and server

+++ mmo/socket.cpp
@@ -144,7 +144,7 @@ int Socket::runClient(){
 
 // Send a client command to the server
 void Socket::sendCommand(std::string msg) {
-    if (send(s, msg.c_str(), msg.length(), 0) < 0)
+    if (send(s, msg.c_str(), (int)msg.length(), 0) < 0)
         std::cout << "Failed to send command: " << msg << std::endl;
     else
         std::cout << "Sent command: " << msg << std::endl;
